refactor(agents): split pilot.ts into modular components (Issue #697)
Split the 895-line `pilot.ts` into smaller, focused modules under
`src/agents/pilot/` directory.

## Changes

| File | Lines | Description |
|------|-------|-------------|
| `pilot.ts` | 646 | Core class (reduced from 895, ~28% reduction) |
| `pilot/types.ts` | 84 | Type definitions (PilotCallbacks, PilotConfig, MessageData) |
| `pilot/message-builder.ts` | 211 | Content building for AI prompts |
| `pilot/index.ts` | 8 | Module exports |

## Key Extractions

### Types (types.ts)
- PilotCallbacks - Platform-specific callback interface
- PilotConfig - Pilot configuration extending BaseAgentConfig
- MessageData - Message processing data structure

### Message Building (message-builder.ts)
- buildEnhancedContent() - Constructs AI prompt with context
- buildToolsSection() - Capability-aware tools documentation
- buildAttachmentsInfo() - File attachment formatting

## Benefits

- ✅ Better separation of concerns
- ✅ Easier testing of message building logic
- ✅ Clearer type definitions
- ✅ Improved maintainability
- ✅ Backward compatibility maintained (types re-exported)

## Test Results

| Metric | Value |
|--------|-------|
| Pilot tests | 12 passed ✅ |
| Total tests | 1556 passed ✅ |
| TypeScript | ✅ No errors |
| (1 pre-existing LOG_LEVEL test failure unrelated to this PR) |

Fixes #697
