+++ src/agents/pilot/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Pilot module exports.
+ *
+ * Issue #697: Extracted from pilot.ts for better organization.
+ */
+
+export type { PilotCallbacks, PilotConfig, MessageData } from './types.js';
+export { buildEnhancedContent, buildToolsSection, buildAttachmentsInfo } from './message-builder.js';
